add check

+++ acc MOCK/back-end/src/main/java/com/sapo/edu/demo/entities/CheckLineEntity.java
@@ -22,10 +22,10 @@ public class CheckLineEntity {
     private String productCode;
     @Column
     @NotNull
-    private Long inventoryQuantity;
+    private Integer inventoryQuantity;
     @Column
     @NotNull
-    private Long actualQuantity;
+    private Integer actualQuantity;
     @Column
     @NotNull
     private String reason;
